+++ source/Engine/InputManager.cpp
@@ -8,8 +8,8 @@ float InputManager::MouseY = 0;
 int InputManager::MouseDown = 0;
 int InputManager::MousePressed = 0;
 int InputManager::MouseReleased = 0;
-float InputManager::MouseMotionX = 0.0f;
-float InputManager::MouseMotionY = 0.0f;
+int InputManager::MouseMotionX = 0;
+int InputManager::MouseMotionY = 0;
 float InputManager::MouseWheelX = 0.0f;
 float InputManager::MouseWheelY = 0.0f;
 int InputManager::MouseMode = MOUSEMODE_DEFAULT;
@@ -366,8 +366,8 @@ void InputManager::RemoveController(int joystickID) {
 }
 
 void InputManager::OnFrameBegin() {
-	MouseMotionX = 0.0f;
-	MouseMotionY = 0.0f;
+	MouseMotionX = 0;
+	MouseMotionY = 0;
 	MouseWheelX = 0.0f;
 	MouseWheelY = 0.0f;
 }
@@ -527,7 +527,7 @@ void InputManager::SetMouseMode(int mode) {
 	MouseMode = mode;
 }
 
-void InputManager::HandleMouseMotionEvent(float motionX, float motionY) {
+void InputManager::HandleMouseMotionEvent(int motionX, int motionY) {
 	MouseMotionX += motionX;
 	MouseMotionY += motionY;
 }
